backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.0
bcrypt<5
PyJWT>=2.8.0
supabase>=1.0.0
youtube-transcript-api>=1.2.4
psycopg2-binary>=2.9.0
google-genai>=1.71.0
httpx>=0.27.0
